[Refactory] 10팀 1차 리펙토링
- 패키지를 크게 domain, glober로 변경
- 패키지명을 소문자로 변경
