--- src/main/java/org/team10/washcode/Enum/UserRole.java
@@ -1,5 +1,0 @@
-package org.team10.washcode.Enum;
-
-public enum UserRole {
-    USER, SHOP, ADMIN
-}
