+++ src/main/java/org/team10/washcode/domain/order/repository/redis/KakaoPaymentInfoRepository.java
@@ -1,0 +1,9 @@
+package org.team10.washcode.domain.order.repository.redis;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+import org.team10.washcode.domain.order.entity.redis.KakaoPaymentInfo;
+
+@Repository
+public interface KakaoPaymentInfoRepository extends CrudRepository<KakaoPaymentInfo, String> {
+}
