+++ src/main/java/org/team10/washcode/domain/order/repository/redis/TokenRepository.java
@@ -1,0 +1,12 @@
+package org.team10.washcode.domain.order.repository.redis;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+import org.team10.washcode.domain.order.entity.redis.Token;
+
+import java.util.Optional;
+
+@Repository
+public interface TokenRepository extends CrudRepository<Token, Long> {
+    Optional<Token> findById(int id);
+}
