--- src/main/java/org/team10/washcode/entity/redis/KakaoPayPgToken.java
@@ -1,15 +1,0 @@
-package org.team10.washcode.entity.redis;
-
-import lombok.AllArgsConstructor;
-import org.springframework.data.redis.core.RedisHash;
-import org.springframework.data.annotation.Id;
-
-@RedisHash("KakaoPayPgToken")
-@AllArgsConstructor
-public class KakaoPayPgToken {
-    @Id
-    private String key;          // Redis Key
-    private Integer partnerOrderId; // Redis Value
-
-    public Integer getPartnerOrderId() { return partnerOrderId; }
-}
