--- src/main/java/org/team10/washcode/entity/redis/KakaoPaymentInfo.java
@@ -1,19 +1,0 @@
-package org.team10.washcode.entity.redis;
-
-import org.springframework.data.annotation.Id;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import org.springframework.data.redis.core.RedisHash;
-
-@RedisHash("KakaoPaymentInfo")
-@AllArgsConstructor
-public class KakaoPaymentInfo {
-    @Id
-    private String key;  // Redis에서 사용할 키 (userId + ":" + cid)
-    private int partnerOrderId;
-    private String tid;
-
-    public Integer getPartnerOrderId() { return partnerOrderId; }
-
-    public String getTid() { return tid; }
-}
