--- src/main/java/org/team10/washcode/repository/redis/TokenRepository.java
@@ -1,12 +1,0 @@
-package org.team10.washcode.repository.redis;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-import org.team10.washcode.entity.redis.Token;
-
-import java.util.Optional;
-
-@Repository
-public interface TokenRepository extends CrudRepository<Token, Long> {
-    Optional<Token> findById(int id);
-}
